+++ tests/apiv2/test_config.py
@@ -16,6 +16,11 @@ from medusa.network_timezones import app_timezone
 from medusa.sbdatetime import date_presets, time_presets
 from medusa.schedulers.utils import all_schedulers
 from medusa.show.recommendations.trakt import TraktPopular
+from medusa.subtitles import (
+    name_from_code,
+    subtitle_code_filter,
+    wanted_languages
+)
 
 from tests.apiv2.conftest import TEST_API_KEY
 
@@ -45,10 +50,7 @@ def config_main(monkeypatch, app_config):
     section_data['wikiUrl'] = app.WIKI_URL
     section_data['donationsUrl'] = app.DONATIONS_URL
     section_data['sourceUrl'] = app.APPLICATION_URL
-    section_data['subtitlesMulti'] = bool(app.SUBTITLES_MULTI)
     section_data['namingForceFolders'] = bool(app.NAMING_FORCE_FOLDERS)
-    section_data['subtitles'] = {}
-    section_data['subtitles']['enabled'] = bool(app.USE_SUBTITLES)
     section_data['recentShows'] = app.SHOWS_RECENT
     section_data['addTitleWithYear'] = bool(app.ADD_TITLE_WITH_YEAR)
     section_data['brokenProviders'] = [provider for provider in app.BROKEN_PROVIDERS if provider]
@@ -197,7 +199,6 @@ async def test_config_get(http_client, create_url, auth_headers, config_main):
 @pytest.mark.gen_test
 @pytest.mark.parametrize('query', [
     'defaultPage',
-    'subtitlesMulti',
     'wikiUrl',
     'sslVerify'
 ])
@@ -370,6 +371,7 @@ def config_system(monkeypatch):
     section_data['appArgs'] = app.MY_ARGS
     section_data['webRoot'] = app.WEB_ROOT
     section_data['runsInDocker'] = bool(app.RUNS_IN_DOCKER)
+    section_data['newestVersionMessage'] = app.NEWEST_VERSION_STRING
     section_data['gitRemoteBranches'] = app.GIT_REMOTE_BRANCHES
     section_data['cpuPresets'] = cpu_presets
 
@@ -863,6 +865,53 @@ async def test_config_get_layout(http_client, create_url, auth_headers, config_l
     expected = config_layout
 
     url = create_url('/config/layout')
+
+    # when
+    response = await http_client.fetch(url, **auth_headers)
+
+    # then
+    assert response.code == 200
+    assert expected == json.loads(response.body)
+
+
+@pytest.fixture
+def config_subtitles():
+    return {
+        'enabled': bool(app.USE_SUBTITLES),
+        'languages': app.SUBTITLES_LANGUAGES,
+        'wantedLanguages': [{'id': code, 'name': name_from_code(code)}
+                            for code in wanted_languages()],
+        'codeFilter': [{'id': code, 'name': name_from_code(code)}
+                       for code in subtitle_code_filter()],
+        'services': app.SUBTITLE_SERVICES,
+        'stopAtFirst': bool(app.SUBTITLES_STOP_AT_FIRST),
+        'eraseCache': bool(app.SUBTITLES_ERASE_CACHE),
+        'location': app.SUBTITLES_DIR,
+        'finderFrequency': int(app.SUBTITLES_FINDER_FREQUENCY),
+        'perfectMatch': bool(app.SUBTITLES_PERFECT_MATCH),
+        'logHistory': bool(app.SUBTITLES_HISTORY),
+        'multiLanguage': bool(app.SUBTITLES_MULTI),
+        'keepOnlyWanted': bool(app.SUBTITLES_KEEP_ONLY_WANTED),
+        'ignoreEmbeddedSubs': bool(app.IGNORE_EMBEDDED_SUBS),
+        'acceptUnknownEmbeddedSubs': bool(app.ACCEPT_UNKNOWN_EMBEDDED_SUBS),
+        'hearingImpaired': bool(app.SUBTITLES_HEARING_IMPAIRED),
+        'preScripts': app.SUBTITLES_PRE_SCRIPTS,
+        'extraScripts': app.SUBTITLES_EXTRA_SCRIPTS,
+        'wikiUrl': app.SUBTITLES_URL,
+        'providerLogins': {
+            'addic7ed': {'user': app.ADDIC7ED_USER, 'pass': app.ADDIC7ED_PASS},
+            'legendastv': {'user': app.LEGENDASTV_USER, 'pass': app.LEGENDASTV_PASS},
+            'opensubtitles': {'user': app.OPENSUBTITLES_USER, 'pass': app.OPENSUBTITLES_PASS}
+        }
+    }
+
+
+@pytest.mark.gen_test
+async def test_config_get_postprocessing(http_client, create_url, auth_headers, config_subtitles):
+    # given
+    expected = config_subtitles
+
+    url = create_url('/config/subtitles')
 
     # when
     response = await http_client.fetch(url, **auth_headers)