+++ themes-default/slim/src/router/routes.js
@@ -54,10 +54,13 @@ const homeRoutes = [
         path: '/home/testRename',
         name: 'testRename',
         meta: {
+            topMenu: 'home',
+            subMenu: showSubMenu,
             title: 'Preview Rename',
             header: 'Preview Rename',
-            topMenu: 'home'
-        }
+            converted: true
+        },
+        component: () => import('../components/test-rename.vue')
     },
     {
         path: '/home/postprocess',
@@ -147,8 +150,10 @@ const configRoutes = [
             title: 'Config - Backup/Restore',
             header: 'Backup/Restore',
             topMenu: 'config',
-            subMenu: configSubMenu
-        }
+            subMenu: configSubMenu,
+            converted: true
+        },
+        component: () => import('../components/config-backup-restore.vue')
     },
     {
         path: '/config/general',
@@ -217,8 +222,10 @@ const configRoutes = [
             title: 'Config - Subtitles',
             header: 'Subtitles',
             topMenu: 'config',
-            subMenu: configSubMenu
-        }
+            subMenu: configSubMenu,
+            converted: true
+        },
+        component: () => import('../components/config-subtitles.vue')
     }
 ];
 
@@ -257,31 +264,6 @@ const addShowRoutes = [
         },
         props: route => ({ ...route.params }),
         component: () => import('../components/new-show.vue')
-    },
-    {
-        path: '/addShows/trendingShows',
-        name: 'addTrendingShows',
-        meta: {
-            topMenu: 'home'
-        }
-    },
-    {
-        path: '/addShows/popularShows',
-        name: 'addPopularShows',
-        meta: {
-            title: 'Popular Shows',
-            header: 'Popular Shows',
-            topMenu: 'home'
-        }
-    },
-    {
-        path: '/addShows/popularAnime',
-        name: 'addPopularAnime',
-        meta: {
-            title: 'Popular Anime Shows',
-            header: 'Popular Anime Shows',
-            topMenu: 'home'
-        }
     }
 ];
 
@@ -354,9 +336,11 @@ const manageRoutes = [
         name: 'manage',
         meta: {
             title: 'Mass Update',
-            header: 'Mass Update',
-            topMenu: 'manage'
-        }
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/manage-mass-update.vue'),
+        props: true
     },
     {
         path: '/manage/backlogOverview',
@@ -364,8 +348,10 @@ const manageRoutes = [
         meta: {
             title: 'Backlog Overview',
             header: 'Backlog Overview',
-            topMenu: 'manage'
-        }
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/manage-backlog.vue')
     },
     {
         path: '/manage/episodeStatuses',
@@ -373,8 +359,10 @@ const manageRoutes = [
         meta: {
             title: 'Episode Overview',
             header: 'Episode Overview',
-            topMenu: 'manage'
-        }
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/manage-episode-status.vue')
     },
     {
         path: '/manage/failedDownloads',
@@ -382,8 +370,10 @@ const manageRoutes = [
         meta: {
             title: 'Failed Downloads',
             header: 'Failed Downloads',
-            topMenu: 'manage'
-        }
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/manage-failed-downloads.vue')
     },
     {
         path: '/manage/manageSearches',
@@ -401,8 +391,11 @@ const manageRoutes = [
         name: 'manageMassEdit',
         meta: {
             title: 'Mass Edit',
-            topMenu: 'manage'
-        }
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/manage-mass-edit.vue'),
+        props: true
     },
     {
         path: '/manage/subtitleMissed',
@@ -410,17 +403,10 @@ const manageRoutes = [
         meta: {
             title: 'Missing Subtitles',
             header: 'Missing Subtitles',
-            topMenu: 'manage'
-        }
-    },
-    {
-        path: '/manage/subtitleMissedPP',
-        name: 'manageSubtitleMissedPP',
-        meta: {
-            title: 'Missing Subtitles in Post-Process folder',
-            header: 'Missing Subtitles in Post-Process folder',
-            topMenu: 'manage'
-        }
+            topMenu: 'manage',
+            converted: true
+        },
+        component: () => import('../components/manage-missing-subtitles.vue')
     }
 ];
 
@@ -432,8 +418,11 @@ const errorLogsRoutes = [
         meta: {
             title: 'Logs & Errors',
             topMenu: 'system',
-            subMenu: errorlogsSubMenu
-        }
+            subMenu: errorlogsSubMenu,
+            converted: true
+        },
+        component: () => import('../components/log-reporter.vue'),
+        props: true
     },
     {
         path: '/errorlogs/viewlog',